pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "fastpdf2png"
version = "1.0.0"
description = "Ultra-fast PDF to PNG converter"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [{ name = "Nataell Cornu" }]
keywords = ["pdf", "png", "converter", "pdfium", "fast", "simd"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

[project.urls]
Homepage = "https://github.com/nataell95/fastpdf2png"
Repository = "https://github.com/nataell95/fastpdf2png"
Issues = "https://github.com/nataell95/fastpdf2png/issues"

[project.optional-dependencies]
pil = ["Pillow"]

[tool.setuptools.packages.find]
include = ["fastpdf2png*"]

[tool.setuptools.package-data]
"fastpdf2png" = ["bin/*"]
